redid bosses and items to be a dictionary using enums as keys to try to fix the intermittent problem. It seems a lot better now but it's still not 100%, but this is a much better way of doing things so I'm gonna keep it.
Basically instead of having a bunch of flags for stuff like showWallJump, I have a flag on Item called saveActiveState which will save the isActive var to UserDefaults when it changes.. The ones that need to be saved (read: walljump, canwalljump, phantoon, eye) are initialized with saveActiveState to true.

Then the menu items and toggles that change it directly modify the Item's isActive state with no intermediary or complicated process.